# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aqarkom API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/auth/ratings": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "operationId": "9d40647a10525a70eab98ca5bdb806eb",
                "requestBody": {
                    "$ref": "#/components/requestBodies/CreateRatingRequest"
                },
                "responses": {
                    "201": {
                        "description": "Rating created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Rating created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/RatingViewModel"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Business logic error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/auth/ratings/{id}": {
            "put": {
                "tags": [
                    "Auth"
                ],
                "operationId": "b3955e97854662285837c7b69a16258a",
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateRatingRequest"
                },
                "responses": {
                    "200": {
                        "description": "Rating updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Rating updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/RatingViewModel"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Rating not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Rating not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Business logic error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Auth"
                ],
                "operationId": "fe08f98367030e93980ed66e7d7ada11",
                "responses": {
                    "200": {
                        "description": "Rating deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Rating deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Rating not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Rating not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Business logic error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/service-providers/{id}": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "operationId": "b061f5f366b04184b076b265cf6a2467",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Service provider ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Service provider retrieved successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UserViewModel"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Service provider not found"
                    },
                    "400": {
                        "description": "Bad Request - User is individual, not service provider"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RatingViewModel": {
                "title": "Rating View Model",
                "description": "Rating data for API responses",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "company_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "rating": {
                        "type": "integer",
                        "maximum": 5,
                        "minimum": 1,
                        "example": 5
                    },
                    "description": {
                        "type": "string",
                        "example": "Great service!",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserViewModel"
                    },
                    "company": {
                        "$ref": "#/components/schemas/UserViewModel"
                    }
                },
                "type": "object"
            },
...
            "UserViewModel": {
                "title": "UserViewModel",
                "description": "User View Model",
                "required": [
                    "id",
                    "email",
                    "phone",
                    "is_active",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string",
                        "nullable": true
                    },
                    "email": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "whatsapp": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string",
                        "nullable": true
                    },
                    "code": {
                        "type": "string",
                        "nullable": true
                    },
                    "is_verified": {
                        "type": "boolean"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "receive_notification": {
                        "type": "boolean"
                    },
                    "receive_messages": {
                        "type": "boolean"
                    },
                    "free_ads": {
                        "type": "integer"
                    },
                    "device_token": {
                        "type": "string",
                        "nullable": true
                    },
                    "last_login": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "is_authentic": {
                        "type": "boolean"
                    },
                    "pending_authentication": {
                        "type": "boolean"
                    },
                    "identity_owner_name": {
                        "type": "string",
                        "nullable": true
                    },
                    "identity_number": {
                        "type": "string",
                        "nullable": true
                    },
                    "commercial_owner_name": {
                        "type": "string",
                        "nullable": true
                    },
                    "commercial_name": {
                        "type": "string",
                        "nullable": true
                    },
                    "commercial_number": {
                        "type": "string",
                        "nullable": true
                    },
                    "commercial_image": {
                        "type": "string",
                        "nullable": true
                    },
                    "identity_image": {
                        "type": "string",
                        "nullable": true
                    },
                    "val_license": {
                        "type": "string",
                        "nullable": true
                    },
                    "transId": {
                        "type": "string",
                        "nullable": true
                    },
                    "requestId": {
                        "type": "string",
                        "nullable": true
                    },
                    "is_nafath_verified": {
                        "type": "boolean"
                    },
                    "about_company": {
                        "type": "string",
                        "nullable": true
                    },
                    "working_hours": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "day": {
                                    "type": "string"
                                },
                                "start_time": {
                                    "type": "string",
                                    "nullable": true
                                },
                                "end_time": {
                                    "type": "string",
                                    "nullable": true
                                },
                                "is_working_day": {
                                    "type": "boolean"
                                }
                            },
                            "type": "object"
                        },
                        "nullable": true
                    },
                    "previous_work_history": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "company_name": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                },
                                "start_date": {
                                    "type": "string",
                                    "format": "date",
                                    "nullable": true
                                },
                                "end_date": {
                                    "type": "string",
                                    "format": "date",
                                    "nullable": true
                                },
                                "is_current_job": {
                                    "type": "boolean"
                                }
                            },
                            "type": "object"
                        },
                        "nullable": true
                    },
                    "services": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "type": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        },
                        "nullable": true
                    },
                    "rating_stats": {
                        "properties": {
                            "average_rating": {
                                "type": "number",
                                "format": "float",
                                "example": 4.5
                            },
                            "total_ratings": {
                                "type": "integer",
                                "example": 10
                            },
                            "ratings_given_count": {
                                "type": "integer",
                                "example": 5
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "ratings_received": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "rating": {
                                    "type": "integer",
                                    "maximum": 5,
                                    "minimum": 1
                                },
                                "description": {
                                    "type": "string",
                                    "nullable": true
                                },
                                "user": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "created_at": {
                                    "type": "string",
                                    "format": "date-time"
                                }
                            },
                            "type": "object"
                        },
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "requestBodies": {
...
            "CreateRatingRequest": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "company_id",
                                "rating"
                            ],
                            "properties": {
                                "company_id": {
                                    "type": "integer",
                                    "example": 2
                                },
                                "rating": {
                                    "type": "integer",
                                    "maximum": 5,
                                    "minimum": 1,
                                    "example": 5
                                },
                                "description": {
                                    "type": "string",
                                    "example": "Great service!",
                                    "nullable": true
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "UpdateRatingRequest": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "rating": {
                                    "type": "integer",
                                    "maximum": 5,
                                    "minimum": 1,
                                    "example": 4
                                },
                                "description": {
                                    "type": "string",
                                    "example": "Updated review",
                                    "nullable": true
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        }
    },
...
}